pyproject.toml
[project]
name = "pomodor-desktop"
version = "0.1.0"
description = "Pomodoro desktop app with local CSV logging"
requires-python = ">=3.11"
dependencies = [
  "pywebview>=5.3",
]
